Novelties:
1,Wireshark (ethereal) dissector plug-in:
I've made a plug-in to decode TFC in wireshark ... and it works! see attached images :)

2, TFC over ESP / TFC over AH:
I was generalizing the code a little bit to be able to put TFC over AH as well. This is useful for development, since we can look inside the packet.
Another way would be to decrypt ESP with wireshark, but I think is harder to set up than to run AH instead.

3, User-Mode Linux (called UML, just like the other one, the moding laguage):
it is a virtual machine technology, running linux over linux. The advantage is that it runs another linux kernel as a user process, so it is easier to
debug. Plus if it freezes, you can simply restart it. And, you can run more of it and set-up a network.

It is easy to compile (ARCH=um), but unfortunately doesn't work with some kernel version combinations for the host and the guest OS. My experimets had
the following outcome:
good host: Fedora Core 5, 2.6.16.19
bad host: Fedora Core 6

good guest: Fedora Core 5, 2.6.18.3
bad guest: 2.6.16.19

Since 2.6.16.19 is the version we are working on, if we want to use UML,  we should patch our sources to a newer version.

4, I was rewriting the queuing and the dequeuing logic:
At queuing time I do NOT modify the packet, just put it in the right queue.
At dequeue time I do the padding and insert the header

5, header insertion:
I've generalized the herder insertion code. It is now capable of inserting the ip_tfc_hdr both as a header and as a footer. Another choice is whether to
add TFC as a next_header to the protocol below (chaining), or leave it  "invisible".
The reason for this is to support some kind of backward compatibility. If you choose "footer" with "invisible" header, we remain compatible with current
IPsec implementations!

6, Algorithm Manager:
I have changed the AM to handle both packet timing and packet size.
It calls the dequeue function at the right time giving also the requested packet size as a parameter

7, fragmentation:
I was rewriting the code a little bit, mostly simplifying it. It seems to work, at least in transport mode...

8, module parameters:
I was adding parameters to the TFC module, this way its functionality can be changed on-the-fly. See images....
Ideally, I think we should have more modules. One giving basic TFC functionality support, and several modules implementing different Algorith Managers.

Nothing else comes to my mind now, although I think I made other changes as well :)

Images:
wireshark_tfc1.png: TFC decoding in action, notice the stack: IP+AH+TFC+ICMP

normal_ping.png: a ping (over AH, though).  122 Byte

ping_tfc.png: ping with TFC. Look at the pad size!

ping_tfc_compat.png: ping with TFC in backward compatibility mode. We still have a footer, even if it is not decoded by wireshark

am_hz.png: changing the speed on the fly with kernel module parameters!

am_pktlen.png: changing packet length. The last one the 2000 is actually 2 packets, fragmented by the underlying IP. See the next image...

ip_fragment.png: This shows the number of packets. You cn clearly see the IP fragmentation of the underlying IP coming in.

ip_fragment_vs_tfc_fragment.png: here first we have pktlen=5000, which means IP fragmentation. Then we have pktlen=50, which means TFC fragmentation.
Look how the first two ICMP datagrams are padded and then fragmented by IP to 4 fragments.
The last ICMP datagrams are intead fragmented by TFC. First we have a truncated ICMP, then a TFC which, according to Wireshark, contains a malformed ICMP
with 0 padding. This is not yet perfect, but as I was writing it down now I've figured out what is the problem :))))

Csaba

git-svn-id: http://minerva.netgroup.uniroma2.it/svn/discreet/tfcproject/trunk/linux-2.6.16.19-TFC@6 0516eb0c-9f27-0410-bdae-aee6bd7d15d1